自分の投稿以外削除不可能にする（仮）

+++ app/src/main/java/com/example/foos/data/domain/DeletePostByPostIdUseCase.kt
@@ -1,12 +1,17 @@
 package com.example.foos.data.domain
 
 import com.example.foos.data.repository.PostsRepository
+import com.google.firebase.auth.FirebaseAuth
 import javax.inject.Inject
 
 class DeletePostByPostIdUseCase @Inject constructor(
-    private val postsRepository: PostsRepository
+    private val postsRepository: PostsRepository,
+    private val firebaseAuth: FirebaseAuth,
 ){
     suspend operator fun invoke(postId: String) {
-        postsRepository.deletePost(postId = postId)
+        val userId = postsRepository.fetchByPostId(postId)?.userId ?: return
+        if (userId == firebaseAuth.currentUser?.uid) {
+            postsRepository.deletePost(postId = postId)
+        }
     }
 }
